refactor: use @ as convention for absolute path, change app's structure of folders
