+++ client/src/app/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { HomePage } from 'components/pages/home';
-import 'scss/style.module.scss';
+import { HomePage } from '@pages/home';
+import '@scss/style.module.scss';
 
 export function App(): JSX.Element {
   return <HomePage />;
