+++ client/src/components/blocks/collectionMovies/cards/index.ts
@@ -1,1 +1,3 @@
 export { Cards } from './Cards';
+
+export type { ICard } from './card';
