+++ client/src/components/blocks/collectionMovies/index.ts
@@ -1,1 +1,3 @@
 export { CollectionMovies } from './CollectionMovies';
+
+export type { ICard } from './cards';
