+++ client/src/components/blocks/index.ts
@@ -1,4 +1,5 @@
 export { CollectionMovies } from './collectionMovies';
+export type { ICard } from './collectionMovies';
 export { Footer } from './footer';
 export { SearchPanel } from './searchPanel';
 export { Logo, Header, MovieForm } from './shared';
