+++ client/src/components/ui/button/Button.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { joinClasses } from 'helpers';
+import { joinClasses } from '@helpers/joinClasses';
 import classes from './Button.module.scss';
 
 export const TYPE_BUTTON_BUTTON = 'button';
