+++ client/src/components/ui/button/buttonClose/ButtonClose.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { joinClasses } from 'helpers';
+import { joinClasses } from '@helpers/joinClasses';
 import classes from './ButtonClose.module.scss';
 
 interface ButtonCloseProps {
