+++ client/src/components/ui/dropdown/index.ts
@@ -1,1 +1,3 @@
 export { Dropdown } from './Dropdown';
+
+export type { IDropdownItem } from './models';
