+++ client/src/components/ui/label/Label.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { joinClasses } from 'helpers';
+import { joinClasses } from '@helpers/joinClasses';
 import classes from './Label.module.scss';
 
 interface LabelProps {
