+++ client/src/components/ui/typography/h4/H4.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { joinClasses } from 'helpers';
+import { joinClasses } from '@helpers/joinClasses';
 import classes from './H4.module.scss';
 
 const CLASSNAME_PRIMARY = 'primary';
