+++ client/src/pages/home/HomePage.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { CollectionMovies, Footer, SearchPanel } from 'components/blocks';
-import { ErrorBoundary } from 'components/errorBoundary';
-import { HorizontalRule } from 'components/ui';
+import { CollectionMovies, Footer, SearchPanel } from '@components/blocks';
+import { ErrorBoundary } from '@components/errorBoundary';
+import { HorizontalRule } from '@components/ui';
 
 export function HomePage(): JSX.Element {
   return (
