+++ client/tsconfig.json
@@ -18,10 +18,16 @@
     "skipLibCheck": true,
     "strict": true,
     "sourceMap": true,
-    "outDir": "./build/",
+    "outDir": "./build",
     "baseUrl": "src",
     "paths": {
-      "src/*": ["src/*"]
+      "@app/*": ["app/*"],
+      "@assets/*": ["assets/*"],
+      "@components/*": ["components/*"],
+      "@helpers/*": ["helpers/*"],
+      "@hooks/*": ["hooks/*"],
+      "@pages/*": ["pages/*"],
+      "@scss/*": ["scss/*"]
     }
   },
   "include": ["src/**/*", "./declaration.d.ts"],
